Casting of external features to float32, fix issue with default lr not being set in params.

+++ data_loader.py
@@ -124,7 +124,7 @@ class ExternalFeature:
         return self.data.shape[1]
 
     def get_feature(self, idx):
-        return torch.tensor(self.data[idx])
+        return torch.tensor(self.data[idx]).float()
 
     @classmethod
     def load_set(cls, feature_loaders, idx):
